+++ terminal-emulator/src/main/rust/tests/content_overflow_test.rs
@@ -1,7 +1,7 @@
 // 测试：验证内容是否正确写入历史缓冲区
 use termux_rust::TerminalEngine;
 
-fn get_row_text(engine: &TerminalEngine, row: i32) -> String {
+fn get_row_text(engine: &TerminalEngine, row: i64) -> String {
     let cols = engine.state.cols as usize;
     let mut text = vec![0u16; cols];
     engine.state.copy_row_text(row, &mut text);
@@ -11,7 +11,7 @@ fn get_row_text(engine: &TerminalEngine, row: i32) -> String {
 #[test]
 fn test_content_goes_beyond_screen() {
     // 创建一个 80x24 的终端，总缓冲区 1000 行
-    let mut engine = TerminalEngine::new(80, 24, 1000, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 1000, 10, 20);
     
     println!("Initial state:");
     println!("  cols={}, rows={}, active_transcript={}", 
@@ -37,7 +37,7 @@ fn test_content_goes_beyond_screen() {
     
     // 检查历史内容（行 -active_transcript 到 -1）
     println!("\nHistory content (sample):");
-    let active = engine.state.main_screen.active_transcript_rows as i32;
+    let active = engine.state.main_screen.active_transcript_rows as i64;
     for i in (-active..0).step_by(10) {
         let text = get_row_text(&engine, i);
         if !text.is_empty() {
@@ -56,7 +56,7 @@ fn test_content_goes_beyond_screen() {
     }
     
     // 验证：最后几行应该在屏幕上
-    // 由于有 77 行历史和 24 行屏幕，Line 100 应该在屏幕底部附近
+    // 由于有 77 行历史 and 24 行屏幕，Line 100 应该在屏幕底部附近
     let mut found_line_100 = false;
     for i in 0..engine.state.rows {
         if get_row_text(&engine, i).contains("Line 100") {
@@ -77,7 +77,7 @@ fn test_content_goes_beyond_screen() {
 #[test]
 fn test_resize_preserves_history() {
     // 创建一个 80x24 的终端
-    let mut engine = TerminalEngine::new(80, 24, 1000, 10, 20);
+    let mut engine = TerminalEngine::new(80 as i64, 24 as i64, 1000, 10, 20);
     
     // 写入 50 行
     for i in 1..=50 {
@@ -88,14 +88,14 @@ fn test_resize_preserves_history() {
     println!("Before resize: active_transcript={}", history_before);
     
     // 缩小到 12 行
-    engine.state.resize(80, 12);
+    engine.state.resize(80 as i64, 12 as i64);
     
     let history_after = engine.state.main_screen.active_transcript_rows;
     println!("After resize to 12 rows: active_transcript={}", history_after);
     
     // 验证历史内容仍然存在
     let mut found_line_1 = false;
-    let active = engine.state.main_screen.active_transcript_rows as i32;
+    let active = engine.state.main_screen.active_transcript_rows as i64;
     for i in -active..engine.state.rows {
         if get_row_text(&engine, i).contains("Line 1") {
             found_line_1 = true;
